Describe artwork instead of returning a path list
Introduce a Artwork property that holds a list of all available artwork and
their types. The paths aren't exported anymore because the GetArtwork() method
is now the prefered way to obtain artwork. The WebArtworkDetailed class is
created by the providers, while only the WebArtwork class is exposed in the
public API.