License all the installers under CPL.
We copy some source code from the WiX source, which is licensed under
the CPL. GPL forbids linking against CPL software (mainly because of
it's choice-of-law clausule), so we can't license parts of the installer
under the GPL. While I'm fine with a more permissive license such as MIT
or BSD for the installers, that only creates more confusion and is a bit
difficult for files that are half authored by Microsoft and half by me.
This also applies to the CustomActions library, as that one is also
linked into the installers.

Note: The CPL applies ONLY TO THE INSTALLERS. The installed software (the
services and WebMediaPortal) is still covered by the GPL.