MAS: Remove category stuff from movies, music and tvshows
We don't implement it yet for the first release, but it won't necessarily be
implemented as user defined categories in the next release. As it will probably
use the already defined categories in the MP plugins, remove the misleading
name.

Also properly implement it for pictures: make the Categories property a list of
WebCategory objects.