Require a call with credentials before giving access to GetMediaItem()
This adds a new method, AuthorizeStreaming(), to the streaming service that is
required to be called to call a method that makes a remote DOS attack possible:
all methods that do either transcoding or stream large files. This AuthorizeStreaming()
method is in the authenticated interface and is implicit with a call to GetServiceDescription()
or InitStream().