WebMP: When seeking in Direct streaming mode, stop previous stream
Currently, when there is being seeked and the Direct streaming mode is used,
the old stream isn't stopped before the new one is started. This caused problems
on systems that were barely powerful enough to handle one stream, but collapse
under the load of two streams. This commit stops the previous stream first, before
starting the new stream.