WSS: Replace all other uses of the ThreadManager
Use either timers, the TPL or unwrapped threads depending on the use case. .NET doesn't
seem to have native support for threads that block reading a stream, so keep using plain
threads for that. We don't have to explicitly kill those on exit though, as they are automatically
killed when the stream is stopped.