WSS: Do not crash on exception in log-parsing threads
With the old ThreadManager, an exception in a thread was caught by some wrapping
code, written to the log and the thread was terminated. Now that we spawn our own
threads in these few places, we've to handle the exceptions by ourselves too.

This has been a regression since e98010ce76c4fe128ec32d6720b1feaacb6f2ddb.