Hosting: Increase all timeouts to 10 hours
This fixes issues with pipe errors after the system has been idle for a
long time. These occurred because WebMP tried to write to the pipe, which
the service had already closed. Increasing the receive timeout should prevent
these errors for idle times lower than 10 hours, but we need a real fix for when
the system is idle longer. Ideally this should be done by a check in MPEServices
or a related class when we try to call a method.