Configuration: Switch to using XmlSerializer
The DataContractSerializer works great, but it has one huge downside: it
requires the nodes in the XML file to be alphabetically sorted. That's just
crazy.

Instead, use the XmlSerializer. This one has a big downside too, but one we can
workaround: by default, it creates assemblies for serializing things on the fly.
However, with sgen.exe we can pre-generate those assemblies. For this, we move
the configuration classes to a separate assembly. With a post-build event we
generate the serialization assembly, which is then used by the Configuration
class to actually perform the serialization. It's a bit of a workaround, but it
seems to work fine.