Configuration: Retry opening a configuration file when reloading
Occasionally Notepad (or another program used to change the configuration files) is still
writing to the file when the Changed event fires. This leads to errors when we try to open
the file, as we don't have access to it. Now we retry opening the file for 5 seconds in case we
can't open it the first time.