Social: Do not pass the IMediaAccessService pipe to the sharing providers
We still have problems with the pipe being broken all the time, which is why we should
refresh the pipe on each call. If we assign the pipe instance to the sharing provider, this
doesn't happen and the sharing provider ends up with a broken pipe. This is complicated
by the fact that we don't do exception-catching in the tasks which call the sharing provider
yet, resulting in a service crash.