WebMP: Skip media items with empty names
Html.ActionLink() throws an exception when you try to create a link with an empty
title, so we should avoid that. While the best solution is to don't return these nonsense
items from the service, that requires all bugs in all plugins to be fixed, so we have to
workaround this.