WebMP: Remove custom HTTP handlers for views, skins and plugins
If we don't touch the default handlers for these files, IIS can handle them without
any intervention from ASP.NET, which is faster. We block access to the cshtml and
executable files with IIS request filtering.

This also avoids the need for skins to place a custom web.config file in their Content
directory to make them accessible.