WSS: Omit Content-Length header for live TV streams (fixes #317)
We can't know the length for live TV streams, since we're reading from a
timeshifting. Unfortunately, GetFileInfo() still returns a size, since the fileinfo
is based upon that of the tsbuffer file. Returning this wrong value (which is
usually just a few hundred bytes) caused problems with clients and browsers,
who disconnected from the stream after reading just a few hundred bytes.

Thanks to CC84 for analysis on this issue.

See the discussion in #394 for more details.