WebMP: Use Proxied mode if requested on an unknown host
When NAT and port forwarding are in use, we can often get requests with a Host
header that we don't know: a domain name that resolves to our external IP, or an
external IP. In these case, it's safe to assume that the Proxied streaming mode is
needed (and even in the rare case that it isn't, we're safe, since Proxied always works).

Generally speaking, this shouldn't change much, since the already existing IsOnLAN()
check should return false for these cases. However, IsOnLAN() isn't the most reliable
method we've implemented and sometimes makes mistakes in complex networks with
multiple network adapters and/or forwarding servers (i.e. port forwarding combined with
a reverse proxy).