Refactored class names from GZIP to Compression
The only GZIP specific code is in the message encoder, I've added
deflate support as well so now the classes have misleading names,
CompressionContext (formerly GzipContext) now carries compression type
information so it actually has something to do in addition to just
existing. It also solves the awkward problem of not knowing if Gzip or
GZip is the correct capitalisation.